Remove RawFile and move Timer and AutoTimer back to glowutils
