+++ source/examples/tessellation/main.cpp
@@ -9,9 +9,6 @@
 #include <glm/gtc/matrix_transform.hpp>
 #include <glm/gtx/transform.hpp>
 
-#include <glowbase/AutoTimer.h>
-#include <glowbase/Timer.h>
-
 #include <glow/Error.h>
 #include <glow/Uniform.h>
 #include <glow/Program.h>
@@ -20,6 +17,7 @@
 #include <glow/logging.h>
 #include <glow/debugmessageoutput.h>
 
+#include <glowutils/Timer.h>
 #include <glowutils/Icosahedron.h>
 #include <glowutils/AdaptiveGrid.h>
 #include <glowutils/Camera.h>
@@ -136,7 +134,7 @@ protected:
     glow::ref_ptr<glowutils::AdaptiveGrid> m_agrid;
 
     glowutils::Camera m_camera;
-    glow::Timer m_time;
+    glowutils::Timer m_time;
 
     vec3 m_rand;
 };
