+++ source/glowutils/CMakeLists.txt
@@ -71,6 +71,7 @@ set(source_path "${CMAKE_CURRENT_SOURCE_DIR}/source")
 set(sources
 	${source_path}/AbstractCoordinateProvider.cpp
 	${source_path}/AdaptiveGrid.cpp
+	${source_path}/AutoTimer.cpp
 	${source_path}/AxisAlignedBoundingBox.cpp
 	${source_path}/AxonometricLookAt.cpp
 	${source_path}/Camera.cpp
@@ -92,6 +93,7 @@ set(sources
 	${source_path}/StackedState.cpp
 	${source_path}/StringSourceDecorator.cpp
 	${source_path}/StringTemplate.cpp
+	${source_path}/Timer.cpp
 	${source_path}/TrackballNavigation.cpp
 	${source_path}/UniformGroup.cpp
 	${source_path}/UnitCube.cpp
@@ -102,6 +104,7 @@ set(sources
 set(api_includes
 	${include_path}/AbstractCoordinateProvider.h
 	${include_path}/AdaptiveGrid.h
+	${include_path}/AutoTimer.h
 	${include_path}/AxisAlignedBoundingBox.h
 	${include_path}/AxonometricLookAt.h
 	${include_path}/Camera.h
@@ -122,6 +125,7 @@ set(api_includes
 	${include_path}/StackedState.h
 	${include_path}/StringSourceDecorator.h
 	${include_path}/StringTemplate.h
+	${include_path}/Timer.h
 	${include_path}/TrackballNavigation.h
 	${include_path}/UniformGroup.h
 	${include_path}/UniformGroup.hpp
